feat: add components for feedback add missing spot feedback
