+++ apps/trading/components/ticket-container/ticket-spot/ticket-limit.tsx
@@ -2,7 +2,6 @@ import BigNumber from 'bignumber.js';
 import uniqueId from 'lodash/uniqueId';
 import { FormProvider, useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { addDays } from 'date-fns';
 
 import { OrderType, OrderTimeInForce } from '@vegaprotocol/types';
 import { useVegaTransactionStore } from '@vegaprotocol/web3';
@@ -12,7 +11,6 @@ import { Form, FormGrid, FormGridCol } from '../elements/form';
 import { type FormFieldsLimit, useLimitSchema } from '../schemas';
 import { TicketTypeSelect } from '../ticket-type-select';
 import { type FormProps } from './ticket';
-import { NON_PERSISTENT_TIF_OPTIONS } from '../constants';
 import { useTicketContext } from '../ticket-context';
 import { SubmitButton } from '../elements/submit-button';
 import { useT } from '../../../lib/use-t';
@@ -41,7 +39,7 @@ export const TicketLimit = (props: FormProps) => {
       type: OrderType.TYPE_LIMIT,
       side: props.side,
       timeInForce: OrderTimeInForce.TIME_IN_FORCE_GTC,
-      expiresAt: addDays(new Date(), 1),
+      expiresAt: undefined,
       postOnly: false,
       reduceOnly: false,
       iceberg: false,
@@ -55,7 +53,7 @@ export const TicketLimit = (props: FormProps) => {
   const tpSl = form.watch('tpSl');
   const iceberg = form.watch('iceberg');
   const tif = form.watch('timeInForce');
-  const isPersistent = !NON_PERSISTENT_TIF_OPTIONS.includes(tif);
+  const isPersistent = utils.isPersistentTif(tif);
 
   return (
     <FormProvider {...form}>
